+++ content_settings/tests/test_all_settings.py
@@ -15,7 +15,7 @@ def test_admin_fields(name):
     value = getattr(content_settings, name)
     setting = ALL[name]
     raw_value = setting.default
-    assert setting.get_admin_preview_value(raw_value, name)
+    assert setting.get_admin_preview_value(raw_value, name) is not None
     assert setting.get_help()